DDBTEAM-399: Only one processor needed
